refactor(desktop): share CommentMarkdown across diff bubble + comment pane
Both the v2 diff comment bubble and the v2 standalone comment pane
were running their own ReactMarkdown setup with subtly different
plugin sets and component maps. Extract the shared rendering into
renderer/components/CommentMarkdown so identical markdown renders
identically across surfaces.

Picks the better of each:
- code: lightweight CommentCodeBlock — Streamdown for mermaid (with
  the project-tuned dark/light theme vars from CommentPane), Prism
  for syntax highlighting, simple <code> for inline. Drops the heavy
  ShowCode wrapper that made small inline review snippets render with
  line numbers + copy button + theme switcher.
- a: target=_blank rel=noopener consistently.
- img: SafeImage everywhere.
- plugins: remarkGfm + rehypeRaw + rehypeSanitize (already aligned).

Per-surface customizations stay local: CommentPane keeps its
CopyableTable as a `components={{ table }}` override; CommentThread
uses the bare base.

v1 CommentPane untouched (sunset path).
