+++ apps/desktop/src/renderer/components/CommentMarkdown/components/CommentCodeBlock/index.ts
@@ -1,0 +1,1 @@
+export { CommentCodeBlock } from "./CommentCodeBlock";
