+++ apps/desktop/src/renderer/components/CommentMarkdown/index.ts
@@ -1,0 +1,1 @@
+export { CommentMarkdown } from "./CommentMarkdown";
