+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/usePaneRegistry/components/CommentPane/CommentPane.tsx
@@ -1,4 +1,3 @@
-import { mermaid } from "@streamdown/mermaid";
 import type { RendererContext } from "@superset/panes";
 import {
 	type ReactNode,
@@ -8,18 +7,8 @@ import {
 	useState,
 } from "react";
 import { LuCheck } from "react-icons/lu";
-import ReactMarkdown from "react-markdown";
-import { Prism as SyntaxHighlighter } from "react-syntax-highlighter";
-import {
-	oneDark,
-	oneLight,
-} from "react-syntax-highlighter/dist/esm/styles/prism";
-import rehypeRaw from "rehype-raw";
-import rehypeSanitize from "rehype-sanitize";
-import remarkGfm from "remark-gfm";
+import { CommentMarkdown } from "renderer/components/CommentMarkdown";
 import { electronTrpcClient } from "renderer/lib/trpc-client";
-import { useTheme } from "renderer/stores";
-import { Streamdown } from "streamdown";
 import type { CommentPaneData, PaneViewerData } from "../../../../types";
 import "./comment-pane.css";
 
@@ -27,115 +16,23 @@ interface CommentPaneProps {
 	context: RendererContext<PaneViewerData>;
 }
 
+const paneComponents = {
+	table: ({ children }: { children?: ReactNode }) => (
+		<CopyableTable>{children}</CopyableTable>
+	),
+};
+
 export function CommentPane({ context }: CommentPaneProps) {
 	const data = context.pane.data as CommentPaneData;
 
 	return (
 		<div className="comment-pane-markdown min-h-0 min-w-0 flex-1 overflow-y-auto select-text">
 			<article className="w-full px-6 py-5">
-				<ReactMarkdown
-					remarkPlugins={[remarkGfm]}
-					rehypePlugins={[rehypeRaw, rehypeSanitize]}
-					components={commentComponents}
-				>
-					{data.body}
-				</ReactMarkdown>
+				<CommentMarkdown body={data.body} components={paneComponents} />
 			</article>
 		</div>
 	);
 }
-
-const mermaidPlugins = { mermaid };
-
-const MERMAID_DARK_VARS = {
-	background: "#1e1e2e",
-	primaryColor: "#313244",
-	primaryTextColor: "#cdd6f4",
-	primaryBorderColor: "#45475a",
-	secondaryColor: "#313244",
-	secondaryTextColor: "#cdd6f4",
-	secondaryBorderColor: "#45475a",
-	tertiaryColor: "#313244",
-	tertiaryTextColor: "#cdd6f4",
-	tertiaryBorderColor: "#45475a",
-	nodeBorder: "#45475a",
-	nodeTextColor: "#cdd6f4",
-	mainBkg: "#313244",
-	clusterBkg: "#1e1e2e",
-	titleColor: "#cdd6f4",
-	edgeLabelBackground: "transparent",
-	lineColor: "#6c7086",
-	textColor: "#cdd6f4",
-};
-
-const MERMAID_LIGHT_VARS = {
-	background: "#ffffff",
-	primaryColor: "#f0f0f4",
-	primaryTextColor: "#1e1e2e",
-	primaryBorderColor: "#d0d0d8",
-	lineColor: "#888",
-	textColor: "#1e1e2e",
-};
-
-function CommentCodeBlock({
-	className,
-	children,
-}: {
-	className?: string;
-	children?: ReactNode;
-}) {
-	const theme = useTheme();
-	const isDark = theme?.type !== "light";
-
-	const match = /language-(\w+)/.exec(className || "");
-	const language = match ? match[1] : undefined;
-	const codeString = String(children).replace(/\n$/, "");
-
-	if (language === "mermaid") {
-		return (
-			<Streamdown
-				mode="static"
-				plugins={mermaidPlugins}
-				mermaid={{
-					config: {
-						theme: "base",
-						themeVariables: isDark ? MERMAID_DARK_VARS : MERMAID_LIGHT_VARS,
-					},
-				}}
-			>
-				{`\`\`\`mermaid\n${codeString}\n\`\`\``}
-			</Streamdown>
-		);
-	}
-
-	if (!language) {
-		return (
-			<code className="px-1.5 py-0.5 rounded bg-muted font-mono text-sm">
-				{children}
-			</code>
-		);
-	}
-
-	return (
-		<SyntaxHighlighter
-			style={
-				(isDark ? oneDark : oneLight) as Record<string, React.CSSProperties>
-			}
-			language={language}
-			PreTag="div"
-			className="rounded-md text-sm"
-		>
-			{codeString}
-		</SyntaxHighlighter>
-	);
-}
-
-const commentComponents = {
-	code: CommentCodeBlock,
-	table: ({ children }: { children?: ReactNode }) => (
-		<CopyableTable>{children}</CopyableTable>
-	),
-};
 
 function CopyableTable({ children }: { children?: ReactNode }) {
 	const tableRef = useRef<HTMLTableElement>(null);
