+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/usePaneRegistry/components/DiffPane/components/CommentThread/comment-thread.css
@@ -153,7 +153,7 @@
 }
 
 .diff-comment-body img,
-.diff-comment-body .diff-comment-img {
+.diff-comment-body .comment-md-img {
 	max-width: 100%;
 	height: auto;
 	border-radius: 4px;
